+++ CHANGELOG.md
@@ -120,3 +120,4 @@
 ## Bugfixes
 
 - [Issue #57](https://github.com/influxdb/influxdb/issues/57). Don't panic when type of time != float
+- [Issue #63](https://github.com/influxdb/influxdb/issues/63). Aggregate queries should not have a sequence_number column